Chore: Resolve random failure with golangci-lint (#21970)
